Chore: remove old merkle_tree
