+++ merkle_tree/Cargo.toml
@@ -3,26 +3,25 @@ name = "olaos_merkle_tree"
 version = "0.1.0"
 edition = "2021"
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
 [dependencies]
-ola_types = { path = "../types" }
+# olavm_core = {package = "core", git = "https://github.com/Sin7Y/olavm.git", branch = "testnet-alpha"}
+olavm_core = { path = "../../olavm/core", package = "core"}
+# olavm_plonky2 = {package = "plonky2", git = "https://github.com/Sin7Y/olavm.git", branch = "testnet-alpha"}
+olavm_plonky2 = { path = "../../olavm/plonky2/plonky2", package = "plonky2" }
 olaos_storage = { path = "../storage", default-features = false }
 ola_utils = { path = "../utils" }
-olaos_logs = { path = "../logs" }
+ola_types = { path = "../types" }
 
-leb128 = "0.2.5"
-once_cell = "1.17.1"
-rayon = "1.3.1"
 thiserror = "1.0"
-tracing = "0.1"
-
-[dev-dependencies]
-assert_matches = "1.5.0"
-clap = { version = "4.2.2", features = ["derive"] }
-insta = { version = "1.29.0", features = ["yaml"] }
-rand = "0.8.5"
-serde = { version = "1", features = ["derive"] }
+web3 = "0.19.0"
+log = "0.4"
+byteorder = "1.3"
 serde_json = "1"
-serde_with = { version = "1", features = ["hex"] }
+serde = { version = "1", features = ["derive"] }
+rayon = "1.7.0"
+rocksdb = { version = "0.21.0", default-features = false, features = ["snappy"] }
+itertools = "0.10"
+bincode = "1.3.3"
 tempfile = "3.0.2"
-test-casing = "0.1.2"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
