+++ merkle_tree/src/iter_ext/merge_join_with_max_predecessor.rs
