+++ merkle_tree/src/patch.rs
