--- merkle_tree/src/types/internal.rs
@@ -1,558 +1,0 @@
-use std::{fmt, num::NonZeroU64};
-
-use crate::{
-    hasher::{HashTree, InternalNodeCache},
-    types::{Key, TreeEntry, ValueHash},
-    utils::SmallMap,
-};
-
-// Size of a (leaf) tree key in bytes.
-pub(crate) const KEY_SIZE: usize = 32;
-// Depth of the tree (= number of bits in `KEY_SIZE`).
-pub(crate) const TREE_DEPTH: usize = KEY_SIZE * 8;
-// Size of a hashed value in bytes.
-pub(crate) const HASH_SIZE: usize = 32;
-
-// Tags associated with a tree.
-#[derive(Debug, Clone)]
-#[cfg_attr(test, derive(PartialEq))]
-pub(crate) struct TreeTags {
-    pub architecture: String,
-    pub depth: usize,
-    pub hasher: String,
-    pub is_recovering: bool,
-}
-
-impl TreeTags {
-    pub const ARCHITECTURE: &'static str = "AR16MT";
-
-    pub fn new(hasher: &dyn HashTree) -> Self {
-        Self {
-            architecture: Self::ARCHITECTURE.to_owned(),
-            hasher: hasher.name().to_owned(),
-            depth: TREE_DEPTH,
-            is_recovering: false,
-        }
-    }
-
-    pub fn assert_consistency(&self, hasher: &dyn HashTree, expecting_recovery: bool) {
-        assert_eq!(
-            self.architecture,
-            Self::ARCHITECTURE,
-            "Unsupported tree architecture `{}`, expected `{}`",
-            self.architecture,
-            Self::ARCHITECTURE
-        );
-        assert_eq!(
-            self.depth, TREE_DEPTH,
-            "Unexpected tree depth: expected {TREE_DEPTH}, got {}",
-            self.depth
-        );
-        assert_eq!(
-            hasher.name(),
-            self.hasher,
-            "Mismatch between the provided tree hasher `{}` and the hasher `{}` used \
-             in the database",
-            hasher.name(),
-            self.hasher
-        );
-
-        if expecting_recovery {
-            assert!(
-                self.is_recovering,
-                "Tree is expected to be in the process of recovery, but it is not"
-            );
-        } else {
-            assert!(
-                !self.is_recovering,
-                "Tree is being recovered; cannot access it until recovery finishes"
-            );
-        }
-    }
-}
-
-// Version-independent information about the tree.
-#[derive(Debug, Default, Clone)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct Manifest {
-    // Number of tree versions stored in the database.
-    pub(crate) version_count: u64,
-    pub(crate) tags: Option<TreeTags>,
-}
-
-impl Manifest {
-    // Returns the version of the tree that is currently being recovered.
-    pub fn recovered_version(&self) -> Option<u64> {
-        if self.tags.as_ref()?.is_recovering {
-            Some(self.version_count.checked_sub(1)?)
-        } else {
-            None
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn new(version_count: u64, hasher: &dyn HashTree) -> Self {
-        Self {
-            version_count,
-            tags: Some(TreeTags::new(hasher)),
-        }
-    }
-}
-
-pub(crate) type NibblesBytes = [u8; KEY_SIZE];
-
-// Unversioned key (a sequence of nibbles) in a radix-16 Merkle tree.
-#[derive(Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub(crate) struct Nibbles {
-    nibble_count: usize,
-    bytes: NibblesBytes,
-}
-
-impl Nibbles {
-    pub const EMPTY: Self = Self {
-        nibble_count: 0,
-        bytes: [0_u8; KEY_SIZE],
-    };
-
-    pub fn nibble(key: &Key, index: usize) -> u8 {
-        const NIBBLES_IN_U64: usize = 16; // 8 bytes * 2 nibbles / byte
-
-        debug_assert!(index < 2 * KEY_SIZE);
-        // Since the `Key` layout is little-endian, we reverse indexing of `u64`:
-        // nibbles 0..=15 are in the final `u64`, etc.
-        let u64_idx = 3 - index / NIBBLES_IN_U64;
-        // The shift in `u64` is reversed as well: the 0th nibble needs the largest shift
-        // `60 = 15 * 4`, the 1st one - 56, etc.
-        let shift_in_u64 = (NIBBLES_IN_U64 - 1 - index % NIBBLES_IN_U64) * 4;
-        ((key.0[u64_idx] >> shift_in_u64) & 0x0f) as u8
-    }
-
-    pub fn new(key: &Key, nibble_count: usize) -> Self {
-        debug_assert!(nibble_count <= 2 * KEY_SIZE);
-        let mut bytes = [0_u8; KEY_SIZE];
-        key.to_big_endian(&mut bytes);
-
-        // Unused bytes (and, if appropriate, the unused nibble in the last used byte)
-        // needs to be zeroized in order for `Ord` / `Eq` / `Hash` traits to work properly.
-        if nibble_count % 2 == 1 {
-            bytes[nibble_count / 2] &= 0xf0;
-        }
-        let meaningful_bytes = (nibble_count + 1) / 2;
-        for byte in bytes.iter_mut().skip(meaningful_bytes) {
-            *byte = 0;
-        }
-
-        Self {
-            nibble_count,
-            bytes,
-        }
-    }
-
-    pub fn from_parts(bytes: NibblesBytes, nibble_count: usize) -> Self {
-        debug_assert!(nibble_count <= 2 * KEY_SIZE);
-        Self {
-            nibble_count,
-            bytes,
-        }
-    }
-
-    pub fn single(nibble: u8) -> Self {
-        assert!(nibble < 16);
-        let mut bytes = NibblesBytes::default();
-        bytes[0] = nibble << 4;
-        Self::from_parts(bytes, 1)
-    }
-
-    pub fn with_version(self, version: u64) -> NodeKey {
-        NodeKey {
-            version,
-            nibbles: self,
-        }
-    }
-
-    pub fn nibble_count(&self) -> usize {
-        self.nibble_count
-    }
-
-    pub fn bytes(&self) -> &NibblesBytes {
-        &self.bytes
-    }
-
-    // Extracts the last nibble and the parent sequence of nibbles
-    // (i.e., one with the last nibble truncated). If this sequence of nibbles is empty,
-    // returns `None`.
-    pub fn split_last(self) -> Option<(Self, u8)> {
-        if self.nibble_count == 0 {
-            return None;
-        }
-
-        let mut truncated_bytes = self.bytes;
-        let last_byte_idx = (self.nibble_count - 1) / 2;
-        let last_byte = self.bytes[last_byte_idx];
-        let last_nibble = if self.nibble_count % 2 == 1 {
-            truncated_bytes[last_byte_idx] = 0;
-            last_byte >> 4
-        } else {
-            truncated_bytes[last_byte_idx] &= 0xf0;
-            last_byte & 15
-        };
-
-        let parent = Self {
-            nibble_count: self.nibble_count - 1,
-            bytes: truncated_bytes,
-        };
-        Some((parent, last_nibble))
-    }
-
-    // Pushes a nibble to the end of this sequence and returns the resulting nibble. Returns `None`
-    // if this nibble is full.
-    pub fn push(self, nibble: u8) -> Option<Self> {
-        if self.nibble_count == KEY_SIZE * 2 {
-            return None;
-        }
-
-        let mut child = self;
-        child.nibble_count += 1;
-        let last_byte_idx = self.nibble_count / 2;
-        if child.nibble_count % 2 == 0 {
-            // The new `nibble` is 4 lower bits
-            child.bytes[last_byte_idx] += nibble;
-        } else {
-            // The new `nibble` is 4 upper bits of a new byte
-            child.bytes[last_byte_idx] = nibble << 4;
-        }
-        Some(child)
-    }
-
-    // Returns nibbles that form a common prefix between these nibbles and the provided `key`.
-    pub fn common_prefix(mut self, other: &Self) -> Self {
-        for i in 0..(self.nibble_count + 1) / 2 {
-            let (this_byte, other_byte) = (self.bytes[i], other.bytes[i]);
-            if this_byte != other_byte {
-                // Check whether the first nibble matches.
-                if this_byte & 0xf0 == other_byte & 0xf0 {
-                    self.nibble_count = i * 2 + 1;
-                    self.bytes[i] &= 0xf0;
-                    self.bytes[(i + 1)..].fill(0);
-                } else {
-                    self.nibble_count = i * 2;
-                    self.bytes[i..].fill(0);
-                }
-                return self;
-            }
-        }
-        self
-    }
-}
-
-impl fmt::Display for Nibbles {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let full_bytes = self.bytes.iter().take(self.nibble_count / 2);
-        for &byte in full_bytes {
-            write!(formatter, "{byte:02x}")?;
-        }
-        if self.nibble_count % 2 == 1 {
-            let last_byte = self.bytes[self.nibble_count / 2];
-            let last_nibble = last_byte >> 4;
-            write!(formatter, "{last_nibble:x}")?;
-        }
-        Ok(())
-    }
-}
-
-impl fmt::Debug for Nibbles {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt::Display::fmt(self, formatter)
-    }
-}
-
-// Versioned key in a radix-16 Merkle tree.
-#[derive(Clone, Copy, PartialEq, Eq, Hash)]
-pub struct NodeKey {
-    pub(crate) version: u64,
-    pub(crate) nibbles: Nibbles,
-}
-
-impl fmt::Debug for NodeKey {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(formatter, "{}:{}", self.version, self.nibbles)
-    }
-}
-
-impl NodeKey {
-    pub(crate) const fn empty(version: u64) -> Self {
-        Self {
-            version,
-            nibbles: Nibbles::EMPTY,
-        }
-    }
-
-    // TODO (BFT-239): Add a fallible version for verifying consistency
-    pub(crate) fn from_db_key(bytes: &[u8]) -> Self {
-        assert!(bytes.len() >= 9, "`NodeKey` is too short");
-
-        let version_bytes: [u8; 8] = bytes[..8].try_into().unwrap();
-        let version = u64::from_be_bytes(version_bytes);
-        let nibble_count = usize::from(bytes[8]);
-        assert!(nibble_count <= 2 * KEY_SIZE);
-        let nibbles_byte_len = (nibble_count + 1) / 2;
-        assert_eq!(nibbles_byte_len, bytes.len() - 9);
-        let mut nibbles = NibblesBytes::default();
-        nibbles[..nibbles_byte_len].copy_from_slice(&bytes[9..]);
-
-        Nibbles::from_parts(nibbles, nibble_count).with_version(version)
-    }
-
-    pub(crate) fn is_empty(&self) -> bool {
-        self.nibbles.nibble_count == 0
-    }
-
-    #[allow(clippy::cast_possible_truncation)]
-    pub(crate) fn to_db_key(self) -> Vec<u8> {
-        let nibbles_byte_len = (self.nibbles.nibble_count + 1) / 2;
-        // ^ equivalent to `ceil(self.nibble_count / 2)`
-        let mut bytes = Vec::with_capacity(9 + nibbles_byte_len);
-        // ^ 8 bytes for `version` + 1 byte for nibble count
-        bytes.extend_from_slice(&self.version.to_be_bytes());
-        bytes.push(self.nibbles.nibble_count as u8);
-        // ^ conversion is safe: `nibble_count <= 64`
-        bytes.extend_from_slice(&self.nibbles.bytes[..nibbles_byte_len]);
-        bytes
-    }
-}
-
-impl fmt::Display for NodeKey {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(formatter, "{}:{}", self.version, self.nibbles)
-    }
-}
-
-// Leaf node of the tree.
-#[derive(Debug, Clone, Copy)]
-#[cfg_attr(test, derive(PartialEq, Eq))]
-pub struct LeafNode {
-    pub(crate) full_key: Key,
-    pub(crate) value_hash: ValueHash,
-    pub(crate) leaf_index: u64,
-}
-
-impl LeafNode {
-    pub(crate) fn new(entry: TreeEntry) -> Self {
-        Self {
-            full_key: entry.key,
-            value_hash: entry.value,
-            leaf_index: entry.leaf_index,
-        }
-    }
-
-    pub(crate) fn update_from(&mut self, entry: TreeEntry) {
-        self.value_hash = entry.value;
-        self.leaf_index = entry.leaf_index;
-    }
-}
-
-// Reference to a child in an [`InternalNode`].
-#[derive(Debug, Clone, Copy)]
-#[cfg_attr(test, derive(PartialEq, Eq))]
-pub(crate) struct ChildRef {
-    pub hash: ValueHash,
-    pub version: u64,
-    pub is_leaf: bool,
-}
-
-impl ChildRef {
-    // Creates a reference to a child with `value_hash` left blank (it will be computed later).
-    pub fn leaf(version: u64) -> Self {
-        Self {
-            hash: ValueHash::default(),
-            version,
-            is_leaf: true,
-        }
-    }
-
-    pub fn internal(version: u64) -> Self {
-        Self {
-            hash: ValueHash::default(),
-            version,
-            is_leaf: false,
-        }
-    }
-}
-
-// Internal node in AR16MT containing up to 16 children.
-#[derive(Default)]
-pub struct InternalNode {
-    children: SmallMap<ChildRef>,
-    cache: Option<Box<InternalNodeCache>>,
-}
-
-impl Clone for InternalNode {
-    fn clone(&self) -> Self {
-        Self {
-            children: self.children.clone(),
-            cache: None,
-            // ^ The cache shouldn't theoretically get invalidated as long as the tree operation
-            // mode doesn't change, but we drop it just to be safe.
-        }
-    }
-}
-
-#[cfg(test)]
-impl PartialEq for InternalNode {
-    fn eq(&self, other: &Self) -> bool {
-        self.children == other.children
-    }
-}
-
-impl fmt::Debug for InternalNode {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut map = formatter.debug_map();
-        for (nibble, child_ref) in self.children() {
-            let nibble = format!("{nibble:x}");
-            map.entry(&nibble, child_ref);
-        }
-        map.finish()
-    }
-}
-
-impl InternalNode {
-    // Number of children in an internal node (= tree radix).
-    pub(crate) const CHILD_COUNT: u8 = 16;
-
-    pub(crate) fn with_capacity(capacity: usize) -> Self {
-        Self {
-            cache: None,
-            children: SmallMap::with_capacity(capacity),
-        }
-    }
-
-    pub(crate) fn child_count(&self) -> usize {
-        self.children.len()
-    }
-
-    pub(crate) fn cache_mut(&mut self) -> Option<&mut InternalNodeCache> {
-        self.cache.as_deref_mut()
-    }
-
-    pub(crate) fn set_cache(&mut self, cache: Box<InternalNodeCache>) -> &mut InternalNodeCache {
-        debug_assert!(self.cache.is_none());
-        self.cache.get_or_insert(cache)
-    }
-
-    pub(crate) fn children(&self) -> impl Iterator<Item = (u8, &ChildRef)> + '_ {
-        self.children.iter()
-    }
-
-    pub(crate) fn child_refs(&self) -> impl Iterator<Item = &ChildRef> + '_ {
-        self.children.values()
-    }
-
-    #[cfg(test)]
-    pub(crate) fn child_refs_mut(&mut self) -> impl Iterator<Item = &mut ChildRef> + '_ {
-        self.children.values_mut()
-    }
-
-    pub(crate) fn last_child_ref(&self) -> (u8, &ChildRef) {
-        self.children.last().unwrap()
-        // ^ `unwrap()` is safe by construction; all persisted internal nodes are not empty
-    }
-
-    pub(crate) fn child_hashes(&self) -> [Option<ValueHash>; Self::CHILD_COUNT as usize] {
-        let mut hashes = [None; Self::CHILD_COUNT as usize];
-        for (nibble, child_ref) in self.children.iter() {
-            hashes[nibble as usize] = Some(child_ref.hash);
-        }
-        hashes
-    }
-
-    pub(crate) fn child_ref(&self, nibble: u8) -> Option<&ChildRef> {
-        self.children.get(nibble)
-    }
-
-    pub(crate) fn child_ref_mut(&mut self, nibble: u8) -> Option<&mut ChildRef> {
-        self.children.get_mut(nibble)
-    }
-
-    pub(crate) fn insert_child_ref(&mut self, nibble: u8, child_ref: ChildRef) {
-        self.children.insert(nibble, child_ref);
-    }
-}
-
-// Tree node (either a leaf or an internal node).
-#[derive(Debug, Clone)]
-#[cfg_attr(test, derive(PartialEq))]
-pub enum Node {
-    // Internal node.
-    Internal(InternalNode),
-    // Tree leaf.
-    Leaf(LeafNode),
-}
-
-impl From<LeafNode> for Node {
-    fn from(leaf: LeafNode) -> Self {
-        Self::Leaf(leaf)
-    }
-}
-
-impl From<InternalNode> for Node {
-    fn from(node: InternalNode) -> Self {
-        Self::Internal(node)
-    }
-}
-
-// Root node of the tree. Besides a [`Node`], contains the general information about the tree
-// (e.g., the number of leaves).
-#[derive(Debug, Clone)]
-#[cfg_attr(test, derive(PartialEq))]
-pub enum Root {
-    // Root for an empty tree.
-    Empty,
-    // Root for a tree with at least one leaf.
-    Filled {
-        // Number of leaves in the tree.
-        leaf_count: NonZeroU64,
-        // Root node of the tree.
-        node: Node,
-    },
-}
-
-impl Root {
-    pub(crate) fn new(leaf_count: u64, node: Node) -> Self {
-        Self::Filled {
-            leaf_count: NonZeroU64::new(leaf_count).unwrap(),
-            node,
-        }
-    }
-
-    pub(crate) fn leaf_count(&self) -> u64 {
-        match self {
-            Self::Empty => 0,
-            Self::Filled { leaf_count, .. } => (*leaf_count).into(),
-        }
-    }
-}
-
-// Stale [`NodeKey`] with information about when it was replaced.
-#[derive(Debug)]
-pub(crate) struct StaleNodeKey {
-    // Replaced key.
-    pub key: NodeKey,
-    // Version of the tree at which it was replaced.
-    pub replaced_in_version: u64,
-}
-
-impl StaleNodeKey {
-    pub fn new(key: NodeKey, replaced_in_version: u64) -> Self {
-        Self {
-            key,
-            replaced_in_version,
-        }
-    }
-
-    pub fn to_db_key(&self) -> Vec<u8> {
-        let mut bytes = Vec::with_capacity(8);
-        bytes.extend_from_slice(&self.replaced_in_version.to_be_bytes());
-        bytes.extend_from_slice(&self.key.to_db_key());
-        bytes
-    }
-}
