--- merkle_tree2/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "merkle_tree2"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-# olavm_core = {package = "core", git = "https://github.com/Sin7Y/olavm.git", branch = "testnet-alpha"}
-olavm_core = { path = "../../olavm/core", package = "core"}
-# olavm_plonky2 = {package = "plonky2", git = "https://github.com/Sin7Y/olavm.git", branch = "testnet-alpha"}
-olavm_plonky2 = { path = "../../olavm/plonky2/plonky2", package = "plonky2" }
-olaos_storage = { path = "../storage", default-features = false }
-ola_utils = { path = "../utils" }
-ola_types = { path = "../types" }
-
-thiserror = "1.0"
-web3 = "0.19.0"
-log = "0.4"
-byteorder = "1.3"
-serde_json = "1"
-serde = { version = "1", features = ["derive"] }
-rayon = "1.7.0"
-rocksdb = { version = "0.21.0", default-features = false, features = ["snappy"] }
-itertools = "0.10"
-bincode = "1.3.3"
-tempfile = "3.0.2"
